Improved keep alive query handling
Judging from the try-with-resources and the small scope, we are not
interested in the ResultSet.

(NB: I originally fixed this on 2.3.x, read that you rather want
fixes on the dev branch, so forward ported it but saw it was
fixed there already, so good on you!
I just propose this as a better fix, do with it what you want)