Fixes #425 a low impact change to satisfy an edge-case.
Still, relying on object identities from getResultSet() is an extremely bad idea, regardless of how some drivers seem to implement it.  Relying on that implementation artifact is contrary to the Statement interface documentation, and likely to fail in a variety of environments.  HikariCP itself does not and will not guarantee the future behaviour of this method.