Add metricsPrefix field to HikariConfig
Adding `metricsPrefix` field to HikariConfig and a default of "pool"
because that's what it is currently. This is work towards #615

Setting the metricsPrefix by default to the poolname + "pool" so that it
matches what's going on in the CodahaleHealthChecker and it will be
easier to update that and keep the old behavior for other people who
don't need to set a custom prefix on there metrics.