Cancel HouseKeeper task on pool shutdown
Periodic house-keeping task is scheduled during construction of
HikariPool. The resulting scheduled task is not stored and therefore it
is not cancelled during pool shutdown.

With standard behaviour, when executor is managed by HikariPool itself,
it is not an issue, because the executor is shut down on pool shutdown.
But when a custom (long running) scheduled executor is supplied via
setScheduledExecutorService, tasks are kept indefinitely causing a
memory leak.